refactor(excursiones): adding more barrel files
