+++ src/components/FormPageLayout/FormPageLayout.test.js
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
-import FormPageLayout from "./index";
+import FormPageLayout from "./FormPageLayout";
 
 // El componente <Link> necesita estar dentro de un Router, por lo que creamos un wrapper.
 const renderWithRouter = (ui, { route = "/" } = {}) => {
