+++ src/components/LoginPage/LoginPageSkeleton.js
@@ -2,7 +2,7 @@ import { Row, Col } from "react-bootstrap";
 import { useSelector } from "react-redux";
 import Skeleton, { SkeletonTheme } from "react-loading-skeleton";
 import "react-loading-skeleton/dist/skeleton.css";
-import FormPageLayout from "../FormPageLayout";
+import FormPageLayout from "../FormPageLayout/FormPageLayout";
 import loginFormStyles from "../LoginForm/LoginForm.module.css";
 
 /** @typedef {import('../../types').RootState} RootState */
