+++ src/components/LoginPage/index.js
@@ -1,5 +1,5 @@
 import LoginForm from "../LoginForm";
-import FormPageLayout from "../FormPageLayout";
+import FormPageLayout from "../FormPageLayout/FormPageLayout";
 import "bootstrap/dist/css/bootstrap.css";
 
 // Componente que representa la página de inicio de sesión.
